pyproject.toml
[project]
name = "job-hunter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "browser-use>=0.12.6",
    "crawl4ai>=0.8.6",
    "duckduckgo-search>=8.1.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.9",
    "langgraph>=1.1.10",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
